+++ apps/builder/app/builder/features/sidebar-left/panels/marketplace/marketplace.tsx
@@ -10,7 +10,7 @@ import type { MarketplaceOverviewItem } from "~/shared/marketplace/types";
 import type { Project } from "@webstudio-is/project";
 import { ExtendedPanel } from "../../shared/extended-panel";
 import { About } from "./about";
-import { trpcClient } from "~/shared/remix/trpc-client";
+import { trpcClient } from "~/shared/trpc/trpc-client";
 
 export const TabContent = ({ onSetActiveTab }: TabContentProps) => {
   const [activeOverviewItem, setAciveOverviewItem] =
