+++ apps/builder/app/services/trcp-router.server.ts
@@ -1,8 +1,10 @@
 import { router } from "@webstudio-is/trpc-interface/index.server";
 import { marketplaceRouter } from "../shared/marketplace/router.server";
+import { domainRouter } from "@webstudio-is/domain/index.server";
 
 export const appRouter = router({
   marketplace: marketplaceRouter,
+  domain: domainRouter,
 });
 
 export type AppRouter = typeof appRouter;
